--- tests/org.eclipse.kura.linux.net.test/src/test/java/org/eclipse/kura/linux/net/iptables/IptablesConfigCoverageTest.java
@@ -1,96 +1,0 @@
- /*******************************************************************************
- * Copyright (c) 2025 Eurotech and/or its affiliates and others
- * 
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- * 
- * SPDX-License-Identifier: EPL-2.0
- * 
- * Contributors:
- *  Eurotech
- *******************************************************************************/
-// Content with portions generated by generative AI platform
-package org.eclipse.kura.linux.net.iptables;
-
-import static org.junit.Assert.fail;
-
-import java.io.File;
-import java.io.FileWriter;
-
-import org.eclipse.kura.KuraException;
-import org.eclipse.kura.KuraIOException;
-import org.eclipse.kura.executor.Command;
-import org.eclipse.kura.executor.CommandStatus;
-import org.eclipse.kura.core.linux.executor.LinuxExitStatus;
-import org.junit.Test;
-
-public class IptablesConfigCoverageTest extends FirewallTestUtils {
-
-    private static class IptablesConfigWithBadTmpPath extends IptablesConfig {
-        IptablesConfigWithBadTmpPath() {
-            super();
-        }
-
-        @Override
-        public String getFirewallConfigTmpFileName() {
-            // Point to a special file path that is not a directory to provoke IOException
-            // when trying to create a file inside it.
-            return "/dev/null/iptables-kura.tmp";
-        }
-    }
-
-    @Test
-    public void saveKuraChains_shouldThrowKuraIOException_onFileCreationFailure() {
-        IptablesConfig cfg = new IptablesConfigWithBadTmpPath();
-        try {
-            cfg.saveKuraChains();
-            fail("Expected KuraIOException to be thrown");
-        } catch (KuraIOException e) {
-            // Expected exception - test passes
-        } catch (KuraException e) {
-            // Check if it's actually a KuraIOException wrapped
-            if (!(e instanceof KuraIOException)) {
-                fail("Expected KuraIOException, but got: " + e.getClass().getSimpleName());
-            }
-        }
-    }
-
-    @Test
-    public void execute_shouldLogDebug_whenDebugEnabled() throws KuraException {
-        // Ensure DEBUG logs are enabled via src/test/resources/log4j.properties
-        setUpMock();
-        IptablesConfig cfg = new IptablesConfig(executorServiceMock);
-        // This will call execute(...) multiple times; with DEBUG enabled, lines 359-361 are executed.
-        cfg.clearAllKuraChains();
-    }
-
-    @Test
-    public void restore_shouldCatchIOException_onDeleteIfExistsFailure() throws Exception {
-        setUpMock();
-
-        // Prepare a non-empty temporary directory so Files.deleteIfExists throws DirectoryNotEmptyException
-        File tempDir = File.createTempFile("iptables-config-test", "dir");
-        // Turn it into a directory: delete the file and create a directory with the same name
-        tempDir.delete();
-        tempDir.mkdir();
-        File inner = new File(tempDir, "inner.txt");
-        try (FileWriter fw = new FileWriter(inner)) {
-            fw.write("data");
-        }
-
-        IptablesConfig cfg = new IptablesConfig(executorServiceMock);
-        // Stub restore command success explicitly to be safe
-        Command cmd = new Command(new String[] { "iptables-restore", tempDir.getAbsolutePath() });
-        cmd.setExecuteInAShell(true);
-        CommandStatus ok = new CommandStatus(cmd, new LinuxExitStatus(0));
-        org.mockito.Mockito.when(executorServiceMock.execute(cmd)).thenReturn(ok);
-
-        // Should not throw; deletion fails inside finally and is caught (line 389)
-        cfg.restore(tempDir.getAbsolutePath());
-
-        // Cleanup: remove inner file then directory
-        inner.delete();
-        tempDir.delete();
-    }
-}
